Fixed the logic on login.
Max, it seems the Realm access control functionality is still confusing for you. I updated this so hopefully it makes it clear.

There is now a method called: updateGlobalRealmPermission which is what we should call only on the first run of the app and with a user that is an admin. The method adds a SyncPermissionChange object to the management Realm for the gloabl Realm URL to set the permissions on that Realm to read/write for all users. In addition, I use the object notifications to listen for changes to the SyncPermissionChange object and notice that it is processed correctly.

You were previously creating a SyncPermissionChange object with a URL: realm://http://138.197.85.79:9080/ and leaving out the path to the global Realm! This results in nothing and appears to actually stop the server from processing any other SyncPermissionChange objects.

David, I am not sure your thesis is correct that this has anything to do with the crash. The crash is occurring in the object store code that processes a change set to produce a notifi
cation. Given that Max was not using a notification block like I am now to listen for change
s to the SyncPermissionChagne object, the crash can't be related. Instead another notificati
on block, such as the one used to update the UI of the converstaion thread might be the culp
rite.

Also, note I introspected my user's management Realm and deleted the old incorrect SyncPermissionChange objects. In doing so, this allowed the server to resume processing the new correct objects based on my updated logic. Bug that needs to be filed with ROS team!